New Anisotropic material + UI portrait support

+++ core/src/primitives/aarect.cpp
@@ -83,11 +83,11 @@ xz_rect::xz_rect(float _x0, float _x1, float _z0, float _z1, float _k, std::shar
 
 bool xz_rect::hit(const ray& r, interval ray_t, hit_record& rec, int depth) const
 {
-    float t = (k - r.origin().y) / r.direction().y;
+    double t = (k - r.origin().y) / r.direction().y;
     if (t < ray_t.min || t > ray_t.max)
         return false;
-    float x = r.origin().x + t * r.direction().x;
-    float z = r.origin().z + t * r.direction().z;
+    double x = r.origin().x + t * r.direction().x;
+    double z = r.origin().z + t * r.direction().z;
     if (x < x0 || x > x1 || z < z0 || z > z1)
         return false;
     //rec.u = (x - x0) / (x1 - x0);
@@ -143,11 +143,11 @@ yz_rect::yz_rect(float _y0, float _y1, float _z0, float _z1, float _k, std::shar
 
 bool yz_rect::hit(const ray& r, interval ray_t, hit_record& rec, int depth) const
 {
-    float t = (k - r.origin().x) / r.direction().x;
+    double t = (k - r.origin().x) / r.direction().x;
     if (t < ray_t.min || t > ray_t.max)
         return false;
-    float y = r.origin().y + t * r.direction().y;
-    float z = r.origin().z + t * r.direction().z;
+    double y = r.origin().y + t * r.direction().y;
+    double z = r.origin().z + t * r.direction().z;
     if (y < y0 || y > y1 || z < z0 || z > z1)
         return false;
     /*rec.u = (y - y0) / (y1 - y0);
